don't delete archive and parent folder after upload

+++ insights/client/phase/v1.py
@@ -186,11 +186,12 @@ def collect_and_output(client, config):
             if config["to_json"]:
                 print(json.dumps(resp))
 
-            # delete the archive
-            if config.keep_archive:
-                logger.info('Insights archive retained in ' + tar_file)
-            else:
-                client.delete_archive(tar_file, delete_parent_dir=True)
+            if not config.payload:
+                # delete the archive
+                if config.keep_archive:
+                    logger.info('Insights archive retained in ' + tar_file)
+                else:
+                    client.delete_archive(tar_file, delete_parent_dir=True)
 
             # if we are rotating the eggs and success on upload do rotation
             try:
